+++ examples/max_LCD/max_LCD.ino
@@ -25,5 +25,5 @@ void loop() {
   // Set the cursor to column 0, line 1 (note: line 1 is the second row, since counting begins with 0):
   lcd.setCursor(0, 1);
   // Print the number of seconds since reset:
-  lcd.print(millis() / 1000);
+  lcd.print((uint32_t)millis() / 1000);
 }
